requirements.txt
streamlit>=1.32.0
faiss-cpu==1.7.4
sentence-transformers==2.5.1
langchain==0.1.12
langchain-community==0.0.29
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0
python-dotenv==1.0.1
groq>=1.0.0
openai>=1.12.0
numpy==1.26.4
scikit-learn==1.4.1.post1
tiktoken==0.6.0
torch==2.2.1+cpu
transformers>=4.38.0
